api_aggregation_service/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn>=0.27.0,<0.28.0
redis>=6.2.0,<6.3.0
motor>=3.7.0,<3.8.0
confluent-kafka>=2.10.0,<2.11.0
python-jose>=3.5.0,<3.6.0
httpx>=0.27.0,<0.28.0
pydantic>=2.6.0,<2.7.0